+++ Plugins/ActorPools/Source/NexusActorPools/Private/NActorPool.cpp
@@ -361,7 +361,7 @@ void FNActorPool::CreateActors(const int32 Count)
 	
 	FActorSpawnParameters SpawnInfo;
 	SpawnInfo.Instigator = nullptr;
-	SpawnInfo.ObjectFlags |= RF_Transient & RF_MarkAsRootSet; // AddToRoot - Pool will manage lifecycle
+	SpawnInfo.ObjectFlags |= RF_Transient | RF_MarkAsRootSet; // AddToRoot - Pool will manage lifecycle
 	SpawnInfo.bDeferConstruction = Settings.HasFlag_DeferConstruction();
 	// We need to tell the spawn to occur and not warn about collisions.
 	SpawnInfo.SpawnCollisionHandlingOverride = ESpawnActorCollisionHandlingMethod::AlwaysSpawn;